+++ package.json
@@ -10,7 +10,7 @@
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "react-scripts": "3.3.0"
+    "react-scripts": "3.4.0"
   },
   "devDependencies": {
     "@mate-academy/eslint-config-react": "*",
