+++ src/components/article-paragraph/article-paragraph.css
@@ -1,0 +1,8 @@
+.article__paragraph {
+  margin: 0;
+  color: #333;
+}
+
+.article__paragraph:not(:first-child) {
+  margin-top: 0.7em;
+}
