+++ src/components/article-paragraph/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './article-paragraph.css';
+
+const ArticleParagraph = ({ text }) => (
+  <p className="article__paragraph">
+    {text}
+  </p>
+);
+
+export default ArticleParagraph;
