+++ src/components/header/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import Navigation from '../navigation';
+
+const Header = () => (
+  <header className="header">
+    <h1 className="header__tmkitle">Site Name</h1>
+    <Navigation />
+  </header>
+);
+
+export default Header;
