+++ src/components/navigation-link/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './navigation-link.css';
+
+const NavigationLink = ({ href = 0, title = 0 }) => (
+  <a className="navigation__link" href={href}>{title}</a>
+);
+
+export default NavigationLink;
