+++ src/components/welcome/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './welcome.css';
+
+const Welcome = () => (
+  <section className="welcome">
+    <span className="welcome__text">Sticky Header!</span>
+  </section>
+);
+
+export default Welcome;
